% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_content}
\alias{get_content}
\title{Title}
\usage{
get_content(ids, search_terms)
}
\arguments{
\item{search_terms}{}
}
\value{

}
\description{
Title
}
\examples{
get_content(c(3487301, 3487301), search_terms = "contraflow")
}
