% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_gazette_feed}
\alias{get_gazette_feed}
\title{Get data from the Gazette API}
\usage{
get_gazette_feed(
  categorycode = 15,
  start_publish_date = "01/01/1998",
  end_publish_date = "31/12/1998",
  base_url = "https://www.thegazette.co.uk/",
  tidy = TRUE
)
}
\arguments{
\item{tidy}{Do you want tidy output? `TRUE` by default}
}
\value{

}
\description{
Get data from the Gazette API
}
\examples{
test_tidy = get_gazette_feed(
  categorycode = 15,
  start_publish_date = "01/01/2021",
  end_publish_date = "31/01/2021"
)
test_jan_2021 = get_gazette_feed(
  categorycode = 15,
  start_publish_date = "01/01/2021",
  end_publish_date = "31/01/2021",
  tidy = FALSE
)
}
