% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_notice_content}
\alias{get_notice_content}
\title{Get notice content}
\usage{
get_notice_content(id, search_terms)
}
\arguments{
\item{search_terms}{}
}
\value{

}
\description{
Get notice content
}
\examples{
# From URL: https://www.thegazette.co.uk/notice/3487301
x4 = get_notice_content(3487301, "contraflow")
x4b = get_notice_content4(3487301, search_terms = c("contraflow", "contra-flow"))
x4c = get_notice_content4(3487301, search_terms = c("cycling"))
x4d = get_notice_content4(3487301, search_terms = c("taxi"))
}
