% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_content}
\alias{get_content}
\title{Get content associated with ids}
\usage{
get_content(ids, search_terms)
}
\arguments{
\item{ids}{Notice IDs}

\item{search_terms}{Character of search terms}
}
\value{

}
\description{
Get content associated with ids
}
\examples{
content = get_content(c(3487301, 3487301), search_terms = "contraflow")
dim(content)
names(content)
content$pub_date
}
